Fix event allocation in queue test
This is a nasty bug and it is more or less guaranteed to happen again,
especially in user code. Let me explain...

In testsuite/queue.f90 we have something like

    program queue
      ! ...
      type(cl_event) :: event

that is, `event' is a static variable for the entire duration of the
program. Hence the compiler wants to be smart and initializes it to zero,
because there is not runtime cost associated with that.

Now the problem is that in src/cl_command_queue_low.c there is a check

    if(*event == NULL){
      // enqueue without setting event
    } else {
      // enqueue with setting event
    }

but since `event' has been initialized to zero we enter the branch that does not
set `event', i.e. we exit the function, expecting that the event has been set,
but in reality `event' still stores 0x0. Thus a subsequent call to
`clWaitForEvents' on `event' fail with error code -58 (CL_INVALID_EVENT).

The workaround in this commit is to make `event' into an array with a single
element. That way `event' must be allocated on the heap[^1] on program start and
will have some uninitialized memory inside that is not 0x0, thereby passing the
test inside `clEnqueueNDRangeKernel_low' and correctly setting the event.

[^1]: This is not necessarily true and might also depend on the compiler. When
optimizing the compiler might still notice that the entire array is static and
small and just push it onto the stack and initialize it to zero, again
triggering the bug. Empirically at least gfortran luckily does not do that.