Fix event allocation in queue test again
Previously, the attempted fix for this problem was to create arrays of events
and hope that the compiled allocates those on the heap with uninitialized memory
that is hopefully random. Unfortunately, this turned out not to be the case,
either because the compiler prefers putting small arrays on the stack and
zero-initializing them, or because the pages on the heap have been zeroed previously.

Hence we take a completely different approach and initialize the events using
explicit bit patterns given by integers of the appropriate size. In the present
case that size is 8 bytes, because cl_event wraps an integer, pointer. Of course
this is very specific to 64-bit platform, but it's seems rather unlikely that
anyone will ever run this test on a 32-bit system.